+++ src/components/DatePicker/DatePicker.tsx
@@ -1,6 +1,7 @@
 import React, { useState, forwardRef } from "react";
 import ReactDatePicker from "react-datepicker";
 
+import { FormRow, FormField } from "../Form/Form";
 import { DatePickerTypes } from "./DatePickerTypes";
 import TextInput, {
   InputProps,
@@ -9,12 +10,15 @@ import TextInput, {
 import generateUUID from "../../helpers/generateUUID";
 import bem from "../../utils/bem";
 
-// Interface used by the `div` or `fieldset` parent wrapper element.
-interface DatePickerWrapperProps {
-  /** ID that other components can cross reference for accessibility purposes. */
-  id?: string;
+interface DateRangeRowProps {
   /** Whether to render a single date input or two for a range of two dates. */
   dateRange?: boolean;
+}
+
+// Interface used by the `div` or `fieldset` parent wrapper element.
+interface DatePickerWrapperProps extends DateRangeRowProps {
+  /** ID that other components can cross reference for accessibility purposes. */
+  id?: string;
   /** Passed to the `TextInput` component to render a label associated with an input field. */
   labelText: string;
   /** Offers the ability to show the label onscreen or hide it. */
@@ -32,6 +36,8 @@ interface CustomTextInputProps extends InputProps {
   onClick?: (data: any) => any;
   /** The ReactDatePicker plugin has its own `required` prop so we use this to pass the value from the parent `DatePicker` component. */
   dsRequired?: boolean;
+  /** Whether or not to display the "Required"/"Optional" text in the label text. */
+  showOptReqLabel?: boolean;
 }
 
 // Main interface for the exported DS DatePicker component.
@@ -50,6 +56,8 @@ export interface DatePickerProps extends DatePickerWrapperProps {
   errored?: boolean;
   /** Adds the 'required' property to the input element(s). */
   required?: boolean;
+  /** Whether or not to display the "Required"/"Optional" text in the label text. */
+  showOptReqLabel?: boolean;
   /** Adds the 'disabled' property to the input element(s). */
   disabled?: boolean;
   /** Modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
@@ -80,6 +88,7 @@ const CustomTextInput = forwardRef<TextInputRefType, CustomTextInputProps>(
       labelText,
       disabled,
       dsRequired,
+      showOptReqLabel,
       errored,
       helperText,
       errorText,
@@ -95,6 +104,7 @@ const CustomTextInput = forwardRef<TextInputRefType, CustomTextInputProps>(
       labelText={labelText}
       disabled={disabled}
       required={dsRequired}
+      showOptReqLabel={showOptReqLabel}
       errored={errored}
       helperText={helperText}
       errorText={errorText}
@@ -130,6 +140,15 @@ const DatePickerWrapper: React.FC<DatePickerWrapperProps> = ({
   );
 
 /**
+ * This is the wrapper component that conditionally renders the input fields
+ * in a `FormRow` only for date range options. This is used for a better visual
+ * layout for the two side-by-side date input fields.
+ * @note This is only used internally for this file.
+ */
+const DateRangeRow: React.FC<DateRangeRowProps> = ({ dateRange, children }) =>
+  dateRange ? <FormRow>{children}</FormRow> : <>{children}</>;
+
+/**
  * Returns a single date input field or two date input fields for a date range.
  */
 function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
@@ -147,6 +166,7 @@ function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
     errorText,
     errored,
     required,
+    showOptReqLabel = true,
     disabled,
     name,
     blockName,
@@ -157,12 +177,15 @@ function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
   const yearsToDisplay = 12;
   // Both ReactDatePicker components share some props.
   let baseDatePickerAttrs = {
+    popperClassName: "datepicker-calendar",
+    popperPlacement: "bottom",
     dateFormat,
     disabled,
   };
   // Both TextInput components share some props.
   let baseCustomTextInputAttrs = {
     dsRequired: required,
+    showOptReqLabel,
     // Always display the labels or the input fields when
     // the DatePicker component is a `dateRange` type.
     showLabel: dateRange ? true : showLabel,
@@ -209,8 +232,9 @@ function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
     endDatePickerAttrs = {
       ...baseDatePickerAttrs,
       selectsEnd: true,
-      startDate: endDate,
       minDate: startDate,
+      startDate,
+      endDate,
     };
     startLabelText = "From";
     endDatePickerElement = (
@@ -252,8 +276,11 @@ function DatePicker(props: React.PropsWithChildren<DatePickerProps>) {
       labelText={labelText}
       className={bem("datePicker", modifiers, blockName, [className])}
     >
-      {startDatePickerElement}
-      {endDatePickerElement}
+      <DateRangeRow dateRange={dateRange}>
+        <FormField>{startDatePickerElement}</FormField>
+
+        {endDatePickerElement && <FormField>{endDatePickerElement}</FormField>}
+      </DateRangeRow>
     </DatePickerWrapper>
   );
 }