+++ CHANGELOG.md
@@ -8,8 +8,14 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Breaking Changes
+
+- Renames original `Card` components to be `CardEdition`.
+
 ### Adds
 
+- Adds new `Card` component.
+- Adds `spacing` prop to `Form` component to allow for spacing variations.
 - Integrates Chakra-UI into the codebase.
 - Exports Chakra's `Box`, `Center`, `Circle`, `Square`, `Stack`, `HStack`, and `VStack` components.
 - Composes the DS `Button` component from Chakra's `Button` component.
@@ -23,6 +29,11 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Exports `VideoPlayerAspectRatios`
 - Adds Tugboat QA configuration for PR preview deployments to replace Netlify.
 - Moved SCSS and CSS variable styles into a custom NYPL theme to extend Chakra's base theme.
+- Updates the `DatePicker`, `TextInput`, `Select`, and `Label` components to pass an optional `showOptReqLabel` prop to conditionally render "Required"/"Optional" in the label text.
+
+### Fixes
+
+- Fixes Accordion FAQ Storybook example rendering bug.
 
 ## 0.23.4
 
@@ -148,7 +159,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Removes `--ui-success` CSS variable and adds `--ui-success-primary` and `--ui-success-secondary`
 - Removes `checkboxId` prop from `Checkbox` component and employs `id` in its place.
 - Removes `defaultChecked` prop from `Checkbox` component.
-- Removed `labelOptions` porp from `Checkbox` component and employs `labelText` in its place.
+- Removed `labelOptions` prop from `Checkbox` component and employs `labelText` in its place.
 
 ### Changes